changed to xConfigDef constructors
