+++ s3-sink-connector/src/main/java/io/aiven/kafka/connect/s3/AivenKafkaConnectS3SinkConnector.java
@@ -21,11 +21,10 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 
+import io.aiven.kafka.connect.s3.config.S3SinkConfigDef;
 import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.connect.connector.Task;
 import org.apache.kafka.connect.sink.SinkConnector;
-
-import io.aiven.kafka.connect.s3.config.S3SinkConfig;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -38,7 +37,7 @@ public class AivenKafkaConnectS3SinkConnector extends SinkConnector {
 
     @Override
     public ConfigDef config() {
-        return S3SinkConfig.configDef();
+        return new S3SinkConfigDef();
     }
 
     @Override
