+++ s3-sink-connector/src/main/java/io/aiven/kafka/connect/s3/config/S3SinkConfig.java
@@ -54,8 +54,6 @@ final public class S3SinkConfig extends S3SinkBaseConfig {
 
     public static final Logger LOGGER = LoggerFactory.getLogger(S3SinkConfig.class);
 
-    private static final String GROUP_AWS = "AWS";
-    private static final String GROUP_FILE = "File";
     // Default values from AWS SDK, since they are hidden
     public static final int AWS_S3_RETRY_BACKOFF_DELAY_MS_DEFAULT = 100;
     public static final int AWS_S3_RETRY_BACKOFF_MAX_DELAY_MS_DEFAULT = 20_000;
@@ -68,68 +66,12 @@ final public class S3SinkConfig extends S3SinkBaseConfig {
     public static final int S3_RETRY_BACKOFF_MAX_RETRIES_DEFAULT = 3;
 
     public S3SinkConfig(final Map<String, String> properties) {
-        super(configDef(), preprocessProperties(properties));
+        super(new S3SinkConfigDef(), preprocessProperties(properties));
     }
 
     static Map<String, String> preprocessProperties(final Map<String, String> properties) {
         // Add other preprocessings when needed here. Mind the order.
         return S3CommonConfig.handleDeprecatedYyyyUppercase(properties);
-    }
-
-    public static ConfigDef configDef() {
-        final var configDef = new S3SinkConfigDef();
-        S3ConfigFragment.update(configDef);
-        addS3partSizeConfig(configDef);
-        FileNameFragment.update(configDef);
-        addOutputFieldsFormatConfigGroup(configDef, null);
-        addDeprecatedTimestampConfig(configDef);
-
-        return configDef;
-    }
-
-    private static void addS3partSizeConfig(final ConfigDef configDef) {
-
-        // add awsS3SinkCounter if more S3 Sink Specific config is added
-        // This is used to set orderInGroup
-        configDef.define(S3ConfigFragment.AWS_S3_PART_SIZE, Type.INT, S3OutputStream.DEFAULT_PART_SIZE,
-                new ConfigDef.Validator() {
-
-                    static final int MAX_BUFFER_SIZE = 2_000_000_000;
-
-                    @Override
-                    public void ensureValid(final String name, final Object value) {
-                        if (value == null) {
-                            throw new ConfigException(name, null, "Part size must be non-null");
-                        }
-                        final var number = (Number) value;
-                        if (number.longValue() <= 0) {
-                            throw new ConfigException(name, value, "Part size must be greater than 0");
-                        }
-                        if (number.longValue() > MAX_BUFFER_SIZE) {
-                            throw new ConfigException(name, value,
-                                    "Part size must be no more: " + MAX_BUFFER_SIZE + " bytes (2GB)");
-                        }
-                    }
-                }, Importance.MEDIUM,
-                "The Part Size in S3 Multi-part Uploads in bytes. Maximum is " + Integer.MAX_VALUE
-                        + " (2GB) and default is " + S3OutputStream.DEFAULT_PART_SIZE + " (5MB)",
-                GROUP_AWS, 0, ConfigDef.Width.NONE, S3ConfigFragment.AWS_S3_PART_SIZE);
-
-    }
-
-    private static void addDeprecatedTimestampConfig(final ConfigDef configDef) {
-        int timestampGroupCounter = 0;
-
-        configDef.define(S3ConfigFragment.TIMESTAMP_TIMEZONE, Type.STRING, ZoneOffset.UTC.toString(),
-                new TimeZoneValidator(), Importance.LOW,
-                "Specifies the timezone in which the dates and time for the timestamp variable will be treated. "
-                        + "Use standard shot and long names. Default is UTC",
-                GROUP_FILE, timestampGroupCounter++, ConfigDef.Width.SHORT, S3ConfigFragment.TIMESTAMP_TIMEZONE);
-
-        configDef.define(S3ConfigFragment.TIMESTAMP_SOURCE, Type.STRING, TimestampSource.Type.WALLCLOCK.name(),
-                new TimestampSourceValidator(), Importance.LOW,
-                "Specifies the the timestamp variable source. Default is wall-clock.", GROUP_FILE,
-                timestampGroupCounter, ConfigDef.Width.SHORT, S3ConfigFragment.TIMESTAMP_SOURCE);
     }
 
     @Override
